moved unzip to implicit class
