+++ src/test/scala/nl.knaw.dans.easy.deposit/servlets/RichZipInputStreamSpec.scala
@@ -13,42 +13,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.deposit
+package nl.knaw.dans.easy.deposit.servlets
 
-import java.io.{ ByteArrayInputStream, InputStream }
+import java.io.ByteArrayInputStream
 import java.nio.charset.StandardCharsets
-import java.nio.file.Paths
 import java.util.zip.ZipInputStream
 
-import better.files.File.temporaryDirectory
-import nl.knaw.dans.bag.v0.DansV0Bag
+import nl.knaw.dans.easy.deposit.TestSupportFixture
 import nl.knaw.dans.easy.deposit.servlets.DepositServlet.BadRequestException
-import nl.knaw.dans.lib.error._
+import resource.managed
 
 import scala.util.Failure
 
-class StagedFilesSpec extends TestSupportFixture {
+class RichZipInputStreamSpec extends TestSupportFixture {
 
   override def beforeEach(): Unit = {
     super.beforeEach()
     clearTestDir()
   }
 
-  private val draftsDir = testDir / "drafts"
-
-  "unzip" should "report invalid content" in {
-    val bag = DansV0Bag
-      .empty(draftsDir / "bag").getOrRecover(e => fail(e.toString, e))
-    bag.save().getOrRecover(e => fail(e.toString, e))
-    val content = "Lorem ipsum est"
-    val fileInBag = "location/in/data/dir/test.txt"
-    val inputStream = new ZipInputStream(new ByteArrayInputStream(content.getBytes(StandardCharsets.UTF_8)))
-
-    temporaryDirectory().apply(
-      StagedFiles(_, bag, Paths.get(fileInBag))
-        .unzip(inputStream) should matchPattern {
+  "extractPlainEntriesTo" should "report invalid content" in {
+    managed(new ZipInputStream(new ByteArrayInputStream("Lorem ipsum est".getBytes(StandardCharsets.UTF_8)))).apply(
+      _.extractPlainEntriesTo(testDir / "staging") should matchPattern {
         case Failure(BadRequestException(s)) if s == "ZIP file is malformed. No entries found." =>
       })
-    inputStream.close()
   }
 }
